feat: update article feature in catalog module

+++ backend/dawn-catalog/src/main/java/com/dawn/catalog/service/Impl/ArticleServiceImpl.java
@@ -6,11 +6,12 @@ import com.dawn.catalog.helper.ArticleMappingHelper;
 import com.dawn.catalog.model.Article;
 import com.dawn.catalog.repository.ArticleRepository;
 import com.dawn.catalog.service.ArticleService;
+import com.dawn.common.core.dto.response.ResponsePage;
 import com.dawn.common.core.exception.wrapper.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,12 +20,10 @@ class ArticleServiceImpl implements ArticleService {
     private final ArticleRepository articleRepository;
 
     @Override
-    public List<ArticleResponse> getAll() {
-        return articleRepository
-                .findAll()
-                .stream()
-                .map(ArticleMappingHelper::map)
-                .toList();
+    public ResponsePage<ArticleResponse> getAll(Pageable pageable) {
+        return ResponsePage.of(articleRepository
+                .findAll(pageable)
+                .map(ArticleMappingHelper::map));
     }
 
     @Override
@@ -36,18 +35,24 @@ class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
+    @Transactional
     public ArticleResponse create(ArticleRequest req) {
         Article article = ArticleMappingHelper.map(req);
+
+        article.setAuthorId(1L);
+        article.setSlug(generateSlug(article.getTitle()));
+
         return ArticleMappingHelper.map(articleRepository.save(article));
     }
 
     @Override
+    @Transactional
     public ArticleResponse update(Long id, ArticleRequest req) {
         Article article = articleRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Article not found"));
+        article.setSlug(generateSlug(article.getTitle()));
         article.setTitle(req.getTitle());
-        article.setSlug(req.getSlug());
         article.setSummary(req.getSummary());
         article.setContent(req.getContent());
         article.setStatus(req.getStatus());
@@ -55,11 +60,16 @@ class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
-    public Void delete(Long id) {
-        articleRepository
-                .findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Article not found"));
+    @Transactional
+    public void delete(Long id) {
+        if (!articleRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Article not found");
+        }
+
         articleRepository.deleteById(id);
-        return null;
+    }
+
+    private String generateSlug(String title) {
+        return title.toLowerCase().replace(" ", "-");
     }
 }
